pyproject.toml
[project]
name = "vynil"
version = "0.1.0"
description = "The Vynil Book Generation Framework"
authors = [
    {name = "Dan Gittik", email = "dan.gittik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "auryn (>=0.6.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "fonttools (>=4.57.0,<5.0.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "latex2mathml (>=3.78.0,<4.0.0)",
    "ebooklib (>=0.19,<0.20)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"
mypy = "^1.15.0"
pytest-cov = "^6.1.1"
poethepoet = "^0.33.1"
types-pygments = "^2.19.0.20250514"
types-pyyaml = "^6.0.12.20250402"

[tool.poe.tasks]
clean = "python dev.py clean"
test = "python dev.py test"
cov = "python dev.py cov"
lint = "python dev.py lint"
type = "python dev.py type"

[tool.poetry.scripts]
vynil = "vynil.cli:cli"

[[tool.mypy.overrides]]
module = ["fontTools.*", "ebooklib.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

